Handle guid_optarg differently if no channel_link
This fixes a bug where the $channel_link would be used as a substring
in forming the $guid_format for some cases of $guid_optarg (namely the
‘auto’ case and the ‘*%s*’ case), **even if $channel_link was empty**
which lead to guid elements that look like this:

    <guid isPermLink="true">/1965-06-19-fox-in-socks.html</guid>

The leading slash probably doesn’t make for a very healthy RSS item.

Of course, this bug only happened if no channel_link was given
(i.e. no ‘-f FORMAT’ ), which was an error in the last major version
anyway (before July 2020). This patch doubles a lot of the code so
it’s not pretty, but basically it omits the guid element altogether if
no absolute $channel_link was provided; in other words, if
$channel_link is empty or is a relative URI, then we can’t correctly
infer the absolute URI so let’s not include it at all.