Clean up last commit
Running a diff between this commit and the commit before last reveals
that not a lot has changed! The last commit was very heavy handed
but by carefully cleaning this up, it's clear that only a subtle change
was needed.

Essentially, we only check for emptiness of $channel_link in the
relative case. This makes sense because we can only infer the
absolute URI from a relative format string if we actually have a non-empty
channel link to work with.